# docs/requirements/api-spec.yaml
openapi: 3.1.0
info:
  title: Staff Procurement Portal API
  version: 0.1.0
  summary: Internal vendor procurement portal for staff
  description: |
    Read-only API backing the Staff Procurement Portal demo. Derived directly
    from `docs/requirements/REQUIREMENTS.md` v0.1 and `docs/requirements/FUNCTIONAL_DESIGN.md`
    v0.1.

    Identity is established by ID.me OAuth 2.0 (authorisation code flow). The backend
    performs the code exchange, validates the ID token, looks the user up in the Oracle
    `STAFF` table, and issues a JWT in an `HttpOnly; SameSite=Lax` cookie named
    `session`. All `/api/vendors/*` endpoints require this cookie. All access to
    Bank details and audit logging are out of scope per ADR-012 — no `BANK_DETAILS` field is returned, and no `AUDIT_LOG` table exists.

    Naming, layer order, and conventions follow `AI_CONTEXT.md`. RBAC follows the
    role/level matrix in `REQUIREMENTS.md` §2.2.
  contact:
    name: Cloud & Things (C&T) — Staff Procurement Portal team
  license:
    name: Proprietary
  x-source-documents:
    - docs/requirements/REQUIREMENTS.md
    - docs/requirements/FUNCTIONAL_DESIGN.md
  x-status: DRAFT
servers:
  - url: http://{ec2_public_ip}
    description: Demo on AWS EC2 (HTTP for the demo; HTTPS is a phase-2 NFR)
    variables:
      ec2_public_ip:
        default: localhost
        description: EC2 public IP supplied at deploy time
  - url: http://localhost
    description: Local Docker Compose development
...
security:
  - cookieAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/auth/callback:
    post:
      tags: [auth]
      summary: Complete ID.me OAuth flow and issue session
      description: |
        Accepts the `code` and `state` returned by ID.me to the SPA's
        `/verification/callback` route. The backend:
          1. Validates `state` against the in-process cache (one-shot, 10 min TTL).
          2. Exchanges `code` for an ID token at ID.me `/oauth/token` (timeout 5 s).
          3. Validates the ID token (`iss`, `aud`, `exp`, `iat`, signature via JWKS).
          4. Looks up the user in Oracle `STAFF` by `EMPLOYEE_ID = sub`.
          5. Applies the access decision tree (see FUNCTIONAL_DESIGN §6.6).
          6. On `GRANTED`, signs a JWT (HS256, 4 h TTL) and sets the `session` cookie.
      operationId: completeLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CallbackRequest" }
      responses:
        "200":
          description: Session established. JWT set as `session` HttpOnly cookie.
          headers:
            Set-Cookie:
              description: |
                `session=<JWT>; HttpOnly; SameSite=Lax; Path=/; Max-Age=14400`
                (Adds `Secure` when HTTPS is enabled.)
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionResponse" }
        "400":
          description: "`state` invalid or expired"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                detail: "Authentication request expired. Please verify again."
        "401":
          description: ID.me token exchange or ID-token validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                detail: "Identity verification failed."
        "403":
          description: |
            User identified but not authorised. The `X-Auth-Reason` response header
            distinguishes denial reasons so the SPA can route to the correct page
            without parsing English copy.
          headers:
            X-Auth-Reason:
              required: true
              schema:
                type: string
                enum: [NON_STAFF, NOT_REGISTERED]
              description: |
                Per [ADR-015](../adr/ADR-015-role-model-simplification.md):
                * `NON_STAFF` — user exists, `STAFF.ROLE = 'NON_STAFF'`. SPA routes to `/access-denied`.
                * `NOT_REGISTERED` — combines `NOT_FOUND`, `IDME_VERIFIED='N'`, `ACTIVE='N'`. Same UX (no account-existence leak).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                non_staff:
                  summary: STAFF.ROLE = 'NON_STAFF'
                  value:
                    detail: "Access denied. Your account does not have portal access."
                not_registered:
                  summary: Not found / not verified / inactive
                  value:
                    detail: "Your identity could not be verified and you do not have access to the portal."
        "502":
          $ref: "#/components/responses/IdentityProviderUnreachable"
...
  /api/auth/dev-login:
    post:
      tags: [auth, dev-only]
      summary: '[DEV ONLY] Sign in as a hardcoded persona — see ADR-014'
      description: |
        **Mounted only when `ENVIRONMENT=dev`.** In any other environment the route
        returns 404 — it is indistinguishable from "endpoint does not exist".

        Five hardcoded personas mint the same session cookie shape as
        `POST /api/auth/callback`. Granted personas return 200 + `Set-Cookie`;
        denied personas return 403 with the same `X-Auth-Reason` headers and
        body copy as the real flow.

        Provided strictly to unblock UI development and stakeholder demos before
        the ID.me sandbox + access service (AC-13) + Oracle seed (AC-14) are
        wired up. See [ADR-014](../adr/ADR-014-demo-persona-login-dev-only.md)
        for kill criteria.
      operationId: devLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DevLoginRequest" }
      responses:
        "200":
          description: Granted persona — session cookie issued (admin_l3, staff_l2, staff_l1).
          headers:
            Set-Cookie:
              description: 'JWT session cookie identical in shape to `/api/auth/callback`.'
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionResponse" }
        "403":
          description: Denied persona (non_staff, not_registered) — matches the `/callback` denial envelope.
          headers:
            X-Auth-Reason:
              description: '`NON_STAFF` or `NOT_REGISTERED` (per ADR-015).'
              schema: { type: string, enum: [NON_STAFF, NOT_REGISTERED] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Endpoint not mounted — `ENVIRONMENT != dev`.
        "422":
          description: Unknown persona key.
  /api/auth/dev-login/available:
    get:
      tags: [auth, dev-only]
      summary: '[DEV ONLY] SPA probe — should the persona panel render?'
      description: |
        Returns 200 only when `ENVIRONMENT=dev`. The SPA renders the persona
        picker on Login when this returns 200 and renders nothing otherwise. No
        other client should depend on this endpoint.
      operationId: devLoginAvailable
      responses:
        "200":
          description: Dev environment — persona login is available.
          content:
            application/json:
              schema:
                type: object
                required: [available]
                properties:
                  available: { type: boolean, enum: [true] }
        "404":
          description: Endpoint not mounted — `ENVIRONMENT != dev`.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        JWT (HS256, claims `sub=staff_id, role, iat, exp, iss, aud`)
        in an `HttpOnly; SameSite=Lax` cookie. TTL configurable via `JWT_TTL_HOURS`
        (default 4). Issued by `POST /api/auth/callback`, cleared by
        `POST /api/auth/logout`. The `session` cookie is the **only** accepted
        credential — no bearer header.
  responses:
...
    IdentityProviderUnreachable:
      description: ID.me is unreachable or returned an unrecoverable error during token exchange.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            detail: "Identity provider unreachable."
  schemas:
...
    CallbackRequest:
      type: object
      required: [code, state]
      additionalProperties: false
      properties:
        code:
          type: string
          minLength: 1
          description: ID.me authorisation code as delivered to `/verification/callback`.
        state:
          type: string
          minLength: 1
          description: ID.me `state` parameter; must match the value cached at `/api/auth/login`.
    DevLoginRequest:
      type: object
      description: |
        Body for `POST /api/auth/dev-login`. Dev-only — see ADR-014. Per
        [ADR-015](../adr/ADR-015-role-model-simplification.md) the persona
        list collapsed from five to four (three business roles plus the
        `not_registered` technical denial path).
      required: [persona]
      additionalProperties: false
      properties:
        persona:
          type: string
          enum: [procurement_supervisor, regular_staff, non_staff, not_registered]
    SessionResponse:
      type: object
      description: |
        Returned by `POST /api/auth/callback` and `GET /api/auth/me`. Per
        [ADR-015](../adr/ADR-015-role-model-simplification.md), `role` is
        the single authority field; `procurement_level` was removed.
      required: [role, staff_id]
      properties:
        role:
          type: string
          enum: [PROCUREMENT_SUPERVISOR, REGULAR_STAFF]
          description: |
            Only the two granted roles appear here. `NON_STAFF` users
            never receive a session — they are denied at
            `/api/auth/callback` with a 403.
        staff_id:
          type: integer
          description: |
            The authenticated user's `STAFF.STAFF_ID`. Used by the SPA to hydrate
            its auth store on every mount (FD §7.2). Included in both
            `POST /api/auth/callback` and `GET /api/auth/me` responses (AC9-D4).

    # ─── Vendors ───────────────────────────────────────────────────────────
...
    Error:
      type: object
      description: |
        FastAPI default error envelope. Browser-facing copy only — stack traces and
        internal codes never leak (NFR-12). Sub-reason for 403s is delivered via the
        `X-Auth-Reason` response header, not in the body.
      required: [detail]
      additionalProperties: false
      properties:
        detail:
          type: string
          description: User-facing error message. Safe to display verbatim.